[KFI]chore(document-viewer-react): move example from src + webpack to ts
Moved example app to example folder to prevent typescript to add example app to dist
Changed
webpack config to typescript config
Fixed material warning in example app
