--- packages/sn-document-viewer-react/index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, height=device-height, minimum-scale=1.0, initial-scale=1.0, user-scalable=0">
-        <title>Document Viewer Example</title>
-        <style>
-            html, body{
-                width: 100%;
-                height: 100%;
-                margin: 0;
-                background-color: #F5F5F5;
-            }
-        </style>
-        <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" />
-    </head>
-    <body>
-        <div id="example" style="width: 100%; height: 100%;"></div>
-
-        <!-- Main -->
-        <script src="assets/bundle.js"></script>
-        <script src="assets/main.bundle.js"></script>
-        <script src="assets/vendors.bundle.js"></script>
-    </body>
-</html>
