+++ packages/sn-document-viewer-react/src/index.ts
@@ -2,4 +2,3 @@ export * from './components'
 export * from './store'
 export * from './models'
 export * from './services'
-export * from './ExampleAppLayout'
