+++ packages/sn-document-viewer-react/tsconfig.json
@@ -3,8 +3,7 @@
   "compilerOptions": {
     "outDir": "dist",
     "composite": true,
-    "jsx": "react",
-    "rootDir": "./src"
+    "jsx": "react"
   },
   "include": ["src"]
 }
